Returned records' attributes are now always vectors, even if only one value was returned.
Changed the function extract-attribute to always return a vector of values.
Before this change, the function would check for the special case of exactly 1 value, in which
case it would just return the value itself. This cause inconsistency in attributes that
are normally arrays (e.g. a person's direct reports) but can sometimes have only a single member
(e.g. a person with only one employee).